OneDrive/Documentos/VPS/AllWinBotSignals/AllWinBotSignals/requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
APScheduler==3.6.3
Babel==2.9.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
matplotlib==3.4.2
matplotlib-inline==0.1.2
mpl-finance==0.10.1
mplfinance==0.12.7a17
mysql-connector-python==8.0.24
psutil==5.8.0
pyTelegramBotAPI==3.7.7
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websocket-client==0.56.0
configparser==5.2.0
woocommerce
